requirements.txt
streamlit
pdfplumber
PyMuPDF
pytesseract
Pillow
requests
google-genai
python-dotenv
pandas